Fix RUF046 lint error in resample_to_spacing
math.ceil() returns an int in Python 3, so the int() wrapper is redundant.
Ruff 0.15.18 (used in CI) flags this under RUF046; an older local ruff did
not have the rule. Verified with `ruff check .` on ruff 0.15.18 and the full
test suite (126 passed).

+++ src/DicomRTTool/Services/StaticScripts.py
@@ -107,7 +107,7 @@ def resample_to_spacing(
     in_size = handle.GetSize()
     in_spacing = handle.GetSpacing()
     new_size = [
-        max(1, int(math.ceil(in_size[i] * in_spacing[i] / output_spacing[i])))
+        max(1, math.ceil(in_size[i] * in_spacing[i] / output_spacing[i]))
         for i in range(3)
     ]
